moved tests from board to player
