+++ src/game.py
@@ -1,11 +1,13 @@
 from src.board import Board
 from src.message import Message as message
+from src.player import Player
 
 
 class Game:
     def __init__(self):
         # initialize a new Board
         self.new_board = Board()
+        self.new_player = Player()
         self.board = self.new_board.starter_board
         self.player_one = self.new_board.player_one
         self.player_two = self.new_board.player_two
@@ -15,7 +17,7 @@ class Game:
         message.display_formatted_board(self)
 
     def get_prompt(self, total_marks_on_board):
-        current_player = self.new_board.get_current_player(total_marks_on_board)
+        current_player = self.new_player.get_current_player(total_marks_on_board)
         if self.new_board.is_full(total_marks_on_board, self.board):
             message.display_game_over_message(self)
         else:
